product state in seperate file
